requirements.txt
Python 3.8+
TensorFlow 2.0+
Keras
NumPy
Matplotlib
Pillow